Deprecation of the Arris Function: I identified that the Arris function in my code was no longer functioning correctly, and I chose to disable it. It's possible that in the future, the function may be reintegrated, but for now, it is marked as deprecated. Users can still use it, but they will be warned about its deprecated status to encourage migration to more suitable alternatives.
Basic Restructuring of Some Activities: Some of the activities in the application were poorly structured or contained redundancies. In this update, I performed a basic restructuring of these activities to enhance code clarity and efficiency.

Change in Function Privacy: To ensure greater security and encapsulation in the code, I changed the visibility level of several functions from "private" to "protected." This helps prevent unauthorized access to certain parts of the code and improves the modularity of the project.

In addition to these major changes, I also made various minor optimizations and bug fixes throughout the code to ensure smoother and more reliable operation of the application. These changes are intended to enhance the maintainability and long-term performance of the app. You can review the complete details of this update in the change log on my GitHub repository.